Refactor: libcrmcommon: separate patchset functions into own file
There are some formatting changes and additional comments but no substantive
changes.

Before:
  4623 lib/common/xml.c

After:
  1748 lib/common/patchset.c
  2969 lib/common/xml.c
  4717 total
